updated notebook RefreshsemanticModel
refreshing the semantic model now is leveraging the sempy function execute_tmsl, this allows using a TMSL script. This script is defined in the variable section and contains the refresh type. The refresh type is now "full."
Still, this is a metadata operation, no data will be "copied" from the delta tables to the semantic model